fix: 修复小数字段不精确的问题，修复cartitem处理错误，移除获取商品详情的店家权限

+++ back-end/Dto/src/main/java/com/xcphoenix/dto/bean/dao/Food.java
@@ -10,6 +10,7 @@ import org.hibernate.validator.constraints.Length;
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 /**
@@ -38,9 +39,9 @@ public class Food {
     private String coverImg;
 
     @DecimalMin(value = "0.0", message = "价格不能为负")
-    private Float originalPrice;
+    private BigDecimal originalPrice;
     @DecimalMin(value = "0.0", message = "价格不能为负")
-    private Float sellingPrice;
+    private BigDecimal sellingPrice;
     private Integer totalSale;
     private Integer monthSale;
